pyproject.toml
[project]
name = "test-cursor"
version = "0.1.0"
description = "Test project for Cursor configuration"
dependencies = [
    "httpx>=0.25.0",
    "requests>=2.31.0",
    "pandas>=2.1.1",
    "rich>=13.6.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
test-network = "tests.network.http_performance:run_tests"
help = "test_cursor.network.http_performance:show_help"

[tool.rye]
managed = true
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/foo"]
